\name{reproducibleCalculationTools-package}
\alias{reproducibleCalculationTools-package}
\alias{reproducibleCalculationTools}
\docType{package}
\title{
reproducibleCalculationTools
}
\description{
Provides utilities to compare subsequent data evaluations. At present, this involves \link{compare_rda_contents}, a function to compare the contents of rda files.
}

\author{
Thomas Braschler\cr\cr

Maintainer: Thomas Braschler <thomas.braschler@gmail.com>

}
\keyword{ package }
